게시글 보기추가 (수정 기능 구현중)

+++ src/main/java/com/ptbh/kyungsunghotel/board/BoardRepository.java
@@ -1,6 +1,9 @@
 package com.ptbh.kyungsunghotel.board;
 
+import com.ptbh.kyungsunghotel.member.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BoardRepository extends JpaRepository<Board, Long> {
+
+public interface BoardRepository extends JpaRepository<Board, Integer> {
+
 }
